Delete interaction.ipynb Add splitted_dataset.ipynb
