+++ .gitignore
@@ -1,0 +1,4 @@
+*/angel.csv
+*/awkward.csv
+*/interaction.ipynb
+*/temp.ipynb
